Fix ambiguous 'else if', adjust array size to account for trailing empty entry
The empty else-if statement was effectively just an else statement;
refactored to match the style of the previous case in the same function.

The array size for tftp_default_options did not account for the last
zeroed entry which was then not created by the compiler. This initially
lead to the problem which was described in [1] and fixed by a bounds check
in the according for loop. Fix the array definition too just to be on the
safe side, in case the array is accessed elsewhere in the same matter.

[1] https://sourceforge.net/p/atftp/bugs/5/