tftpd.c: Only drop privs if requested or running as root & check for failure
Previously, the return code of setuid and setgid wasn't checked at all,
which their according man pages point out as "a grave security error".
However, after adding the checks atftpd would fail when run as non-root, since
obviously a setuid to nobody would fail. So we only try to drop privs now
if we're either running as root or if explicitly requested via --user
and/or --group.